# perf(common): improve performance of col-aware encoding for structs (serialization part)