--- a/src/common/src/util/value_encoding/column_aware_row_encoding.rs
+++ b/src/common/src/util/value_encoding/column_aware_row_encoding.rs
@@ -26,12 +26,13 @@ use ahash::HashMap;
 use bitfield_struct::bitfield;
 use bytes::{Buf, BufMut};
 use rw_iter_util::ZipEqDebug;
+use smallvec::{SmallVec, smallvec};
 
 use super::error::ValueEncodingError;
 use super::{Result, ValueRowDeserializer, ValueRowSerializer};
 use crate::catalog::ColumnId;
 use crate::row::Row;
-use crate::types::{DataType, Datum, ScalarRefImpl, ToDatumRef};
+use crate::types::{DataType, Datum, ScalarRefImpl, StructType, ToDatumRef};
 use crate::util::value_encoding as plain;
 
 /// Serialize and deserialize functions that recursively use [`ColumnAwareSerde`] for nested fields
@@ -65,13 +66,10 @@ mod new_serde {
     //
     // Recursively construct a new column-aware `Serializer` for nested fields.
     fn new_serialize_struct(struct_type: &StructType, value: StructRef<'_>, buf: &mut impl BufMut) {
-        let serializer = super::Serializer::new(
-            &struct_type.ids().unwrap().collect_vec(), // TODO: avoid this clone
-            struct_type.types().cloned(),              // TODO: avoid this clone
-        );
+        let serializer = super::Serializer::from_struct(struct_type.clone()); // cloning `StructType` is lightweight
 
         // `StructRef` is interpreted as a `Row` here.
-        let bytes = serializer.serialize(value); // TODO: serialize into the buf directly
+        let bytes = serializer.serialize(value); // TODO: serialize into the buf directly if we can reserve space accurately
         buf.put_u32_le(bytes.len() as _);
         buf.put_slice(&bytes);
     }
@@ -252,8 +250,8 @@ struct Header {
 /// `RowEncoding` holds row-specific information for Column-Aware Encoding
 struct RowEncoding {
     header: Header,
-    offsets: Vec<u8>,
-    buf: Vec<u8>,
+    offsets: SmallVec<[u8; 16]>,
+    data: Vec<u8>,
 }
 
 /// A trait unifying [`ToDatumRef`] and already encoded bytes.
@@ -290,8 +288,8 @@ impl RowEncoding {
     fn new() -> Self {
         RowEncoding {
             header: Header::new(),
-            offsets: vec![],
-            buf: vec![],
+            offsets: Default::default(),
+            data: Default::default(),
         }
     }
 
@@ -313,76 +311,119 @@ impl RowEncoding {
         self.header.set_offset(offset_width);
 
         self.offsets
-            .reserve_exact(usize_offsets.len() * offset_width.width());
+            .resize(usize_offsets.len() * offset_width.width(), 0);
+
+        let mut offsets_buf = &mut self.offsets[..];
         for &offset in usize_offsets {
-            self.offsets
-                .put_uint_le(offset as u64, offset_width.width());
+            offsets_buf.put_uint_le(offset as u64, offset_width.width());
         }
     }
 
-    fn encode<T: Encode>(&mut self, datums: impl IntoIterator<Item = T>, data_types: &[DataType]) {
+    fn encode<T: Encode>(
+        &mut self,
+        datums: impl IntoIterator<Item = T>,
+        data_types: impl IntoIterator<Item = &DataType>,
+    ) {
         debug_assert!(
-            self.buf.is_empty(),
+            self.data.is_empty(),
             "should not encode one RowEncoding object multiple times."
         );
         let datums = datums.into_iter();
         let mut offset_usize = Vec::with_capacity(datums.size_hint().0);
         for (datum, data_type) in datums.zip_eq_debug(data_types) {
-            offset_usize.push(self.buf.len());
-            datum.encode_to(data_type, &mut self.buf);
+            offset_usize.push(self.data.len());
+            datum.encode_to(data_type, &mut self.data);
         }
         self.set_offsets(&offset_usize);
     }
 }
 
+mod data_types {
+    use crate::types::{DataType, StructType};
+
+    pub trait DataTypes: Clone {
+        fn my_iter(&self) -> impl Iterator<Item = &DataType>;
+    }
+
+    impl DataTypes for Vec<DataType> {
+        fn my_iter(&self) -> impl Iterator<Item = &DataType> {
+            self.iter()
+        }
+    }
+
+    impl DataTypes for StructType {
+        fn my_iter(&self) -> impl Iterator<Item = &DataType> {
+            self.types()
+        }
+    }
+}
+use data_types::DataTypes;
+
 /// Column-Aware `Serializer` holds schema related information, and shall be
 /// created again once the schema changes
 #[derive(Clone)]
-pub struct Serializer {
-    encoded_column_ids: Vec<u8>,
-    data_types: Vec<DataType>,
+pub struct Serializer<D: DataTypes = Vec<DataType>> {
+    encoded_column_ids: EncodedColumnIds,
+    data_types: D,
+}
+
+type EncodedColumnIds = SmallVec<[u8; 32]>;
+
+fn encode_column_ids(column_ids: impl ExactSizeIterator<Item = ColumnId>) -> EncodedColumnIds {
+    // currently we hard-code ColumnId as i32
+    let mut encoded_column_ids = smallvec![0; column_ids.len() * 4];
+    let mut buf = &mut encoded_column_ids[..];
+    for id in column_ids {
+        buf.put_i32_le(id.get_id());
+    }
+    encoded_column_ids
 }
 
 impl Serializer {
     /// Create a new `Serializer` with given `column_ids` and `data_types`.
     pub fn new(column_ids: &[ColumnId], data_types: impl IntoIterator<Item = DataType>) -> Self {
-        // currently we hard-code ColumnId as i32
-        let mut encoded_column_ids = Vec::with_capacity(column_ids.len() * 4);
-        for id in column_ids {
-            encoded_column_ids.put_i32_le(id.get_id());
-        }
-
         Self {
-            encoded_column_ids,
+            encoded_column_ids: encode_column_ids(column_ids.iter().copied()),
             data_types: data_types.into_iter().collect(),
         }
     }
+}
 
+impl Serializer<StructType> {
+    pub fn from_struct(struct_type: StructType) -> Self {
+        Self {
+            encoded_column_ids: encode_column_ids(struct_type.ids().unwrap()),
+            data_types: struct_type,
+        }
+    }
+}
+
+impl<D: DataTypes> Serializer<D> {
     fn datum_num(&self) -> usize {
         self.encoded_column_ids.len() / 4
     }
 
     fn serialize_raw<T: Encode>(&self, datums: impl IntoIterator<Item = T>) -> Vec<u8> {
         let mut encoding = RowEncoding::new();
-        encoding.encode(datums, &self.data_types);
+        encoding.encode(datums, self.data_types.my_iter());
         self.finish(encoding)
     }
 
     fn finish(&self, encoding: RowEncoding) -> Vec<u8> {
         let mut row_bytes = Vec::with_capacity(
-            5 + self.encoded_column_ids.len() + encoding.offsets.len() + encoding.buf.len(), /* 5 comes from u8+u32 */
+            5 + self.encoded_column_ids.len() + encoding.offsets.len() + encoding.data.len(), /* 5 comes from u8+u32 */
         );
         row_bytes.put_u8(encoding.header.into_bits());
         row_bytes.put_u32_le(self.datum_num() as u32);
         row_bytes.extend(&self.encoded_column_ids);
         row_bytes.extend(&encoding.offsets);
-        row_bytes.extend(&encoding.buf);
+        row_bytes.extend(&encoding.data);
 
         row_bytes
     }
 }
 
-impl ValueRowSerializer for Serializer {
+impl<D: DataTypes> ValueRowSerializer for Serializer<D> {
     /// Serialize a row under the schema of the Serializer
     fn serialize(&self, row: impl Row) -> Vec<u8> {
         assert_eq!(row.len(), self.datum_num());